Joao

+++ src/main/java/br/edu/ifsp/domain/usecases/ativo/acao/ExcluirAcaoUseCase.java
@@ -3,6 +3,7 @@ package br.edu.ifsp.domain.usecases.ativo.acao;
 import br.edu.ifsp.domain.entities.ativo.Acao;
 import br.edu.ifsp.domain.entities.log.LogAtivo;
 import br.edu.ifsp.domain.entities.log.LogAtivoEnum;
+import br.edu.ifsp.domain.usecases.grupo.GrupoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.LogAtivoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.SalvarHistoricoAtivoUseCase;
 import br.edu.ifsp.domain.usecases.utils.Notification;
@@ -11,10 +12,12 @@ import br.edu.ifsp.domain.usecases.utils.Validator;
 public class ExcluirAcaoUseCase {
     private AcaoDAO acaoDAO;
     private LogAtivoDAO logAtivoDAO;
+    private GrupoDAO grupoDAO;
 
-    public ExcluirAcaoUseCase(AcaoDAO acaoDAO, LogAtivoDAO logAtivoDAO) {
+    public ExcluirAcaoUseCase(AcaoDAO acaoDAO, LogAtivoDAO logAtivoDAO, GrupoDAO grupoDAO) {
         this.acaoDAO = acaoDAO;
         this.logAtivoDAO = logAtivoDAO;
+        this.grupoDAO = grupoDAO;
     }
 
     public boolean delete(Acao acao) {
@@ -23,6 +26,10 @@ public class ExcluirAcaoUseCase {
 
         if(notif.hasErrors()) {
             throw new IllegalArgumentException(notif.errorMessage());
+        }
+
+        if(this.grupoDAO.findOneByAtivo(acao).isPresent()) {
+            throw new IllegalArgumentException("Cant delete ativo thats in a grupo");
         }
 
         boolean flag = this.acaoDAO.delete(acao);
