+++ src/main/java/br/edu/ifsp/domain/usecases/ativo/rendafixa/ExcluirRendaFixaUseCase.java
@@ -3,6 +3,7 @@ package br.edu.ifsp.domain.usecases.ativo.rendafixa;
 import br.edu.ifsp.domain.entities.ativo.RendaFixa;
 import br.edu.ifsp.domain.entities.log.LogAtivo;
 import br.edu.ifsp.domain.entities.log.LogAtivoEnum;
+import br.edu.ifsp.domain.usecases.grupo.GrupoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.LogAtivoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.SalvarHistoricoAtivoUseCase;
 import br.edu.ifsp.domain.usecases.utils.Notification;
@@ -11,10 +12,12 @@ import br.edu.ifsp.domain.usecases.utils.Validator;
 public class ExcluirRendaFixaUseCase {
     private RendaFixaDAO rendaFixaDAO;
     private LogAtivoDAO logAtivoDAO;
+    private GrupoDAO grupoDAO;
 
-    public ExcluirRendaFixaUseCase(RendaFixaDAO rendaFixaDAO, LogAtivoDAO logAtivoDAO) {
+    public ExcluirRendaFixaUseCase(RendaFixaDAO rendaFixaDAO, LogAtivoDAO logAtivoDAO, GrupoDAO grupoDAO) {
         this.rendaFixaDAO = rendaFixaDAO;
         this.logAtivoDAO = logAtivoDAO;
+        this.grupoDAO = grupoDAO;
     }
 
     public boolean delete(RendaFixa rendaFixa) {
@@ -23,6 +26,10 @@ public class ExcluirRendaFixaUseCase {
 
         if(notif.hasErrors()) {
             throw new IllegalArgumentException(notif.errorMessage());
+        }
+
+        if(this.grupoDAO.findOneByAtivo(rendaFixa).isPresent()) {
+            throw new IllegalArgumentException("Cant delete ativo thats in a grupo");
         }
 
         boolean flag = this.rendaFixaDAO.delete(rendaFixa);
